// src/services/rest-api/customer-service/model/financial-info-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FinancialInfoModel
 */
export interface FinancialInfoModel {
    /**
     * 
     * @type {Array<string>}
     * @memberof FinancialInfoModel
     */
    'investmentPurposeCode'?: Array<string> | null;
    /**
     * 
     * @type {string}
     * @memberof FinancialInfoModel
     */
    'investmentFundCode'?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof FinancialInfoModel
     */
    'incomeSourceCode'?: Array<string> | null;
    /**
     * 
     * @type {string}
     * @memberof FinancialInfoModel
     */
    'incomeRateCode'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof FinancialInfoModel
     */
    'investmentYear'?: number;
    /**
     * 
     * @type {number}
     * @memberof FinancialInfoModel
     */
    'assetValue'?: number;
}