// src/services/rest-api/customer-service/models/FinancialInfoRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.332
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FinancialInfoRequest
 */
export interface FinancialInfoRequest {
    /**
     * 
     * @type {Array<string>}
     * @memberof FinancialInfoRequest
     */
    investmentPurposeCode?: Array<string> | null;
    /**
     * 
     * @type {string}
     * @memberof FinancialInfoRequest
     */
    investmentFundCode?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FinancialInfoRequest
     */
    investmentPurposeOther?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof FinancialInfoRequest
     */
    incomeSourceCode?: Array<string> | null;
    /**
     * 
     * @type {string}
     * @memberof FinancialInfoRequest
     */
    incomeRateCode?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FinancialInfoRequest
     */
    incomeSourceOther?: string | null;
    /**
     * 
     * @type {number}
     * @memberof FinancialInfoRequest
     */
    investmentYear?: number;
    /**
     * 
     * @type {number}
     * @memberof FinancialInfoRequest
     */
    assetValue?: number;
}

/**
 * Check if a given object implements the FinancialInfoRequest interface.
 */
export function instanceOfFinancialInfoRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function FinancialInfoRequestFromJSON(json: any): FinancialInfoRequest {
    return FinancialInfoRequestFromJSONTyped(json, false);
}

export function FinancialInfoRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): FinancialInfoRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'investmentPurposeCode': !exists(json, 'investmentPurposeCode') ? undefined : json['investmentPurposeCode'],
        'investmentFundCode': !exists(json, 'investmentFundCode') ? undefined : json['investmentFundCode'],
        'investmentPurposeOther': !exists(json, 'investmentPurposeOther') ? undefined : json['investmentPurposeOther'],
        'incomeSourceCode': !exists(json, 'incomeSourceCode') ? undefined : json['incomeSourceCode'],
        'incomeRateCode': !exists(json, 'incomeRateCode') ? undefined : json['incomeRateCode'],
        'incomeSourceOther': !exists(json, 'incomeSourceOther') ? undefined : json['incomeSourceOther'],
        'investmentYear': !exists(json, 'investmentYear') ? undefined : json['investmentYear'],
        'assetValue': !exists(json, 'assetValue') ? undefined : json['assetValue'],
    };
}

export function FinancialInfoRequestToJSON(value?: FinancialInfoRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'investmentPurposeCode': value.investmentPurposeCode,
        'investmentFundCode': value.investmentFundCode,
        'investmentPurposeOther': value.investmentPurposeOther,
        'incomeSourceCode': value.incomeSourceCode,
        'incomeRateCode': value.incomeRateCode,
        'incomeSourceOther': value.incomeSourceOther,
        'investmentYear': value.investmentYear,
        'assetValue': value.assetValue,
    };
}